+++ app/src/components/Main/index.js
@@ -1,9 +1,5 @@
 import React from "react";
 import useStyles from "./styles.js";
-import {
-    Typography,
-} from '@material-ui/core'
-
 import Sidebar from '../Sidebar';
 
 export default function Main({children}) {
