+++ app/src/contexts/AuthContext.js
@@ -1,0 +1,2 @@
+import { createContext } from 'react'
+export const AuthContext = createContext();
