+++ app/src/pages/Login/styles.js
@@ -20,6 +20,7 @@ const useStyles = makeStyles((theme) => ({
         boxSizing: 'border-box',
         width: 405,
         padding:'48px 72px 0px 72px',
+        boxShadow: "0px 0px 32px rgba(0, 0, 0, 0.1)",
         borderRadius: 16,
         marginBottom: 32,
         "& > div":{
